Update webhook parsing to new QuickNode format and bigger refactor (#24)
QuickNode webhooks changed their payload structure. While updating the
app to work with the new structure, I decided to improve the
architecture of the entire repo to be more generelizable to make it
easier to add new events in the future.

Once v3 comes around we'll want to monitor more events and
could then simply extend this repo to add new events